refactor: refactor board.service
