+++ src/resources/boards/board.service/index.js
@@ -1,0 +1,1 @@
+module.exports = require('./board.service')
